refactor(public-form/tests): shards tests out also
